requirements.txt
flask
pySerial
PyVirtualSerial